refactor: Controller 제거. Application으로 통합.
- Application 이 컨트롤러의 역할을 한다.
